refactor: DiaryCard props를 diary 객체로 통합 (#19)
