+++ src/entities/diary/index.ts
@@ -1,4 +1,4 @@
-export type { Diary, EmotionIntensity, TodayDiary } from "./model/diary.types";
+export type { Diary, DiaryDetail, EmotionIntensity, TodayDiary } from "./model/diary.types";
 export { DiaryCard } from "./ui/DiaryCard";
 export { DiaryChip } from "./ui/DiaryChip";
 export { DiaryListItem } from "./ui/DiaryListItem";
